refactor ranking service into scoring module
