+++ app/routes/_actions+/reply.tsx
@@ -5,7 +5,10 @@ import { db } from '#app/db.ts'
 import { fallacyDetection } from '#app/modules/fallacies/fallacy-detection-client.ts'
 import { storeFallacies } from '#app/modules/fallacies/fallacy-repository.ts'
 import { createPost } from '#app/modules/posts/post-service.ts'
-import { getCommentTreeState, getReplyTree } from '#app/repositories/ranking.ts'
+import {
+	getCommentTreeState,
+	getReplyTree,
+} from '#app/modules/scoring/ranking-service.ts'
 import { requireUserId } from '#app/utils/auth.server.ts'
 
 type ReplyData = {
