+++ app/routes/discussions.tsx
@@ -4,7 +4,7 @@ import moment from 'moment'
 import { PostContent } from '#app/components/building-blocks/post-content.tsx'
 import { Markdown } from '#app/components/markdown.tsx'
 import { db } from '#app/db.ts'
-import * as rankingTs from '#app/repositories/ranking.ts'
+import { getChronologicalToplevelPosts } from '#app/modules/scoring/ranking-service.ts'
 import { type FrontPagePost } from '#app/types/api-types.ts'
 import { getUserId } from '#app/utils/auth.server.ts'
 
@@ -12,7 +12,7 @@ export async function loader({ request }: LoaderFunctionArgs) {
 	const userId: string | null = await getUserId(request)
 	const loggedIn = userId !== null
 	const feed = await db.transaction().execute(async trx => {
-		return await rankingTs.getChronologicalToplevelPosts(trx)
+		return await getChronologicalToplevelPosts(trx)
 	})
 	return { loggedIn, feed }
 }
